requirements.txt
numpy==1.16.4
Pillow==6.1.0
tqdm==4.32.1
matplotlib==3.1.0
pydot==1.4.1
Keras==2.2.4
tensorflow==1.13.2